+++ src/pkg_chk/main.cxx
@@ -66,7 +66,7 @@ namespace {
             std::vector<std::string> argv = {pkgxx::shell, "-s", "--"};
             argv.insert(argv.end(), args.begin(), args.end());
             pkgxx::harness prog(
-                pkgxx::shell, argv, cwd, env_mod,
+                pkgxx::shell, argv, cwd, env_mod, std::nullopt,
                 pkgxx::harness::fd_action::pipe,
                 pkgxx::harness::fd_action::pipe,
                 pkgxx::harness::fd_action::merge_with_stdout);